Umbau der Funktion ListIngredience

+++ app/src/main/java/com/example/cocktailmachine/ui/ListOfIngredience/RecyclerViewListenerListIngredience.java
@@ -2,5 +2,5 @@ package com.example.cocktailmachine.ui.ListOfIngredience;
 
 public interface RecyclerViewListenerListIngredience {
 
-    void selectIngredience(int Id);
+    void selectIngredience(long Id);
 }
